requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pinecone>=6.0.0
sentence-transformers>=3.0.0
anthropic>=0.28.0
praw>=7.7.1
youtube-transcript-api>=0.6.2
FlagEmbedding>=1.2.10
transformers>=4.30.0,<5.0.0
slowapi>=0.1.9
redis>=5.0.0
pydantic-settings>=2.2.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
nltk>=3.8.1
tavily-python>=0.5.0
